organize-bookmarks/pyproject.toml
[project]
name = "organize-bookmarks"
version = "0.1.0"
description = ""
authors = [
    {name = "Nate Stott", email = "nate.stott@pm.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.93.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
# disable “installing” your own code as a package
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
